Bug #1150157: kernel/gcov/gcc_3_4.c: use struct_size() in kzalloc()
[root cause  ] clang gcov support needed
[changes     ] patch from linux master
[side effects] no
[self test   ]: yes
[download normally]: yes
[self test   ]: yes
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: s9863a1h10
[test case]: clang gcov profiling
[reviewers   ] nianfu.bai
[change_type ] feature
[tag_product ] common

This patch comes from linux mater branch.

----------------------------------

kernel/gcov/gcc_3_4.c: use struct_size() in kzalloc()

One of the more common cases of allocation size calculations is finding
the size of a structure that has a zero-sized array at the end, along
with memory for some number of elements for that array.  For example:

  struct foo {
      int stuff;
      void *entry[];
  };

  instance = kzalloc(sizeof(struct foo) + sizeof(void *) * count, GFP_KERNEL);

Instead of leaving these open-coded and prone to type mistakes, we can
now use the new struct_size() helper:

  instance = kzalloc(struct_size(instance, entry, count), GFP_KERNEL);

This code was detected with the help of Coccinelle.

Change-Id: I72db5ec890c39f9880581a38e303b2bed8f5b6fc
Link: http://lkml.kernel.org/r/20190109172445.GA15908@embeddedor
Signed-off-by: Gustavo A. R. Silva <gustavo@embeddedor.com>
Reviewed-by: Peter Oberparleiter <oberpar@linux.ibm.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Candle Sun <candle.sun@unisoc.com>