Bug #1008857 sched: Add account_wait_time interface into schedtune
[root cause  ] runnable time for tasks is not considered as busy in WALT
[changes     ] add account_wait_timer interface into schedtune
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:yes
[board]:sharkL5
[test case]: check scheduling behaviour
[reviewers   ]
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

walt_account_wait_time in WALT is set as false by default in the
previous commit, which prohibit some little tasks that runnable
for a long time going to big core. However, runnable time is
accounted as busy time is helpful for performance.

Add account_wait_time interface into schedtune, which can control
runnable time of tasks is busy or not in different groups as needed.

Change-Id: Id36b52e75ef7707c29062091d59dd519c3e2a26f
Signed-off-by: Ke Wang <ke.wang@spreadtrum.com>