Bug #1020675 usb: musb: Fix USB card unidentified issue when plugin microUSB interface
[root cause  ] When we first pull some USB card readers out, sometimes CSR registers
		will not be clear so as to controller status is unnormal on next plug in.
[changes     ] Clear TX/RX CSR registers when we pull any USB interfaces out.
[side effects] NA
[self test   ]: yes
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: sharkle
[test case]: build & plug in/out
[reviewers   ] jiayong.yang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

From Customer's tester report, some USB cards could not be identified
was likely to happen. As we analysed, when we first pull USB card readers
out CSR registers will not be clear at once. It would be likely to lead
to keeping the value of TX/RX CSR registers until next pulg in so as to
enumulation failed.

Our solution is that clear TX/RX CSR registers when we pull any USB
interfaces out.

Change-Id: Ib86dd09dff1fccee34595f74ca3f754204209494
Signed-off-by: Joey Wu <joey.wu@unisoc.com>