Bug #941064 arm64/arm: dts: Modify software delay time for dvfs
[root cause  ] The software delay time is too long
[changes     ] Decrease the software delay time acoording to the hardware latency
[side effects] no
[self test   ]: dvfs test
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:no
[is there dependence]:no
[confirm dependent commit]:no
[board]:ums312_1h10 ums312-1h10-go ums312-2h10-go
[test case]:dvfs test,reboot test,sleep wakeup
[reviewers   ] jack.liu ling_ling.xu jingchao.ye sheng.zhu
[change_type ] bringup --> sprd_adaptation
[tag_product ] common

1)Because there is not the problem of competing the channel with others when we use adi channel
to adjust the voltage for little cores, and the governor can guarantee the period of frequency modulation
is longer than the maximum latency that hardware spends in accomplishing to adjust cpu cores's frequency,
so we decrease the software delay time for little core cluster to zero.

2)Because there is the problem of competing the channel with other drivers when we use i2c channel to
adjust the voltage for big core, so the dvfs driver needs to lock the i2c channel for a while until finish
adjusting frequency by hardware, in addition the governor can guarantee the period of frequency modulation
is longer than the maximum latency that hardware spends in accomplishing to adjust cpu cores's frequency,
at present the maximum latency measured is about 250us,so we decrease the software delay time for
big cores and cluster to 250us.

Change-Id: I38dc0b4848e9a7481fcd7709ad1a3a1aa3c1af8a
Signed-off-by: Jack Liu <jack.liu@spreadtrum.com>