Bug #1017538 pci: sprd: enable global slave error response
[root cause  ] bring up
[changes     ] pcie driver
[side effects] none
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: roc1
[test case]: boot ok
[reviewers   ] billows.wu; jason.wu
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

If RC send some commands to access some memory addresses of EP
side that can not be accessed, these commands cannot be completed
and will be blocked in the EP's receive buffer. When EP's credit
is exhausted, the axi slave interface will be hanged if there are
continuous requests. Enable the global slave error response to
notify the CPU that an exception has occurred.

Change-Id: Id5bd7d38bdf727cc3035414457664e8e022e2feb
Signed-off-by: Jolly Li <jolly.li@spreadtrum.com>