Bug #1038423 power: supply: sc2703: Fix kernel panic issue when phone reboot test
[root cause  ] fix kernel panic issue when phone reboot test
[changes     ] fix kernel panic issue when phone reboot test
[side effects] no
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: Sharkl5
[test case]: compile ok, charger ok, reboot ok
[reviewers   ] baolin.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

When the phone boot process, becauese the 2703 charger depend on phy
registration, so if the phy resource is not ready, the 2703 charger will
always probe, and the bus regulator will register repeatedly but in this
moment the debtry->d_inode may be 0, and currently the musb will access
the vbus regulator interface, so will led kernel panic. Now we register
vbus regulator after 2703 charger get phy handle.

Change-Id: I8b19fbc2fc221f4edc8452151a7a83faeef1c3bb
Signed-off-by: Yuanjiang Yu <yuanjiang.yu@spreadtrum.com>