Bug #1044326 DIRTY: arm64: dts: Modify PCIe node order
[root cause  ] Orca PCIe should be init first
[changes     ] dts
[side effects] no
[self test   ]: yes
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: no
[is there dependence]: no
[confirm dependent commit]: yes
[board]: ud710_2h10
[test case]: orca pcie works well
[reviewers   ] Jason Wu
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

ROC1 has two pcie controller, one is for marlin3 whose start address
is 0x20c00000, one is for orca whose start address is 0x20d00000.
If marlin3 is not opened, orca pcie domain is 0000. However, if
marlin3 is enabled, the pcie controller for marlin3 will be initialized
first. The pcie domain of marlin3 will be 0000 and the pcie domain of
orca will be changed to 0001.

Orca EP must deal with the problem about pcie domain changing. However,
it needs some times. This patch can solve this problem quickly.
So we need this patch until orca EP can deal with the chang of pcie
domain automatically.

Change-Id: If5b8e3021d961faeeb64281ea3c6e2beb571f800
Signed-off-by: Billows Wu <Billows.Wu@spreadtrum.com>