Bug #1021417 i2c: sprd: Modify the hardware i2c clear ACK and FIFO policy
[root cause  ] Modify the hardware i2c clear ACK and FIFO policy.
[changes     ] i2c-sprd-hw.c
[side effects]no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:yes
[confirm dependent commit]:yes
[board]:ums312-1h10
[test case]:test the i2c functions.
[reviewers   ] lanqing.liu
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Since the hardware i2c may have a scenario where the software channel and
the hardware channel simultaneously transmit data, if the software channel
clears the FIFO after the data is transmitted, and the hardware is transmitting
data at this time, the data of the hardware channel is also cleared, resulting
in abnormal communication. In addition, because the existing ASIC design directly
writes the ACK bit, the entire controller will be reset, If the hardware channel
is still communicating at this time, the data will be lost. Therefore, it is
necessary to first determine whether the communication is abnormal, and then
decide whether to clear the FIFO and ACK.

Change-Id: Ie45a9d65bcc88e7b4ba6e728147fbbb7fb3855f4
Signed-off-by: Xingxing Luo <xingxing.luo@unisoc.com>