Bug #1021417 i2c: sprd: Change the way to check data transfer is completed
[root cause  ] Change the way to check data transfer is completed.
[changes     ] i2c-sprd-hw.c
[side effects]no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:yes
[confirm dependent commit]:yes
[board]:ums312-1h10
[test case]:test the i2c functions.
[reviewers   ] lanqing.liu
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

The scene where the software channel and the hardware channel transmit
data at the same time, no matter which one is completed, the irq will be
set bit,If irq is set bit, it is impossible to distinguish which one set
irq. Therefore, irq cannot be cleared in the software. At the same time,
the ASIC design of the write channel is that the data in the FIFO will be
pulled low and the pending bit will only remain about 114ns. The CPU
cannot capture the pending state, so it takes 2ms to wait in the
software. This time can ensure that the data has been transmitted.
After 2ms, the pending bit is judged. If there is no pending, the
transmission is completed. If there is pending, then judge. Is the
software channel or hardware channel exception, and do the corresponding
processing.

Change-Id: I5c7aa8460541892804d81dad5aa64a56cae427f3
Signed-off-by: Xingxing Luo <xingxing.luo@unisoc.com>