Bug #1026922 usb: musb: Fix usb suspend fail
[root cause  ] fix usb suspend fail
[changes     ] musb_core.c
[side effects] none
[self test   ]:usb test
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:roc1
[test case]:usb test
[reviewers   ] jiayong.yang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

The musb devctl session bit is set until vbus is invlid
in unisoc platform. So the usb core can not enter suspend
when we plug out usb, because musb check session bit is set
and don't put block runtime PM in disconnect interrupt.

Besides, we use typec to notify usb plug in/out event,
and need not to check session bit.

So we remove the musb_pm_runtime_check_session function.

Change-Id: I060fd0f995b7d7eecd47fdf9efa0977bebc96923
Signed-off-by: Hongbo Li <hongbo.li@unisoc.com>