Bug #1049258 cpufreq: schedutil: Add DTS parsing logic for slack timer
[root cause  ]sharkL5 has auto scaling dvfs for hardware
[changes     ]remove the software slack timer
[side effects]no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:yes
[board]:sharkL5
[test case]:check the slack timer whether starting
[reviewers   ]
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Before sharkL5, there is no cpu dvfs scaling for hardware after entering
WFI with a higher OPP. So, a slack timer was added in schedutil to check
whether cpu entered WFI with a higher OPP and will scale to the lowest
OPP for cpu.

From sharkL5, hardware will auto scale cpu dvfs to the lowest OPP under
WFI and there is no need to do anything for software.

Add DTS parsing logic in schedutil and find the key string of
'sugov_slack_timer' to decide whether start the slack timer.

Change-Id: I16303ea93b306dc98ac3fdea3b4db85e9785976f
Signed-off-by: Ke Wang <ke.wang@spreadtrum.com>