Bug #1047335 arm64: dts: increase r61350 vbp+vsync to 34
[root cause  ] dpu underflow
[changes     ] panel driver
[side effects] none
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:ums312
[test case]:call test ok
[reviewers   ] leon.he thomas.chai
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

The internal line buffer is 32 lines, which can maintain 32 lines of time.
The time for filling data can't exceed 32 lines.
If the HW finds more than enough, it will report underflow.

For normal display, each buffer fill time will be within 32 lines.
The initial fill time is limited by (Vsync + VBP). hope >=32 lines time.
If it is too small, it's easy to have undeflow at the top of the frame.

Change-Id: Id6f395ed8863c1c2dd745576976f854116b8bae0
Signed-off-by: Kevin Tang <kevin.tang@spreadtrum.com>