Bug #1053539 power: supply: charger-manager: Optimize charging timeout strategy
[root cause  ] optimize charging timeout strategy
[changes     ] optimize charging timeout strategy
[side effects] no
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: Roc1
[test case]: compile ok
[reviewers   ] baolin.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

When running the monkey test, the system power sonsumption very large,
so the charging current into the battery is very small. In this scenes,
the charging duration is more than 6 hours, and meet the charging timeout
mechanism, it will cause low power shutdown issue. Now we optimize
charging timemout strategy, if the  charging time is more than 6 hours
and the open circuit voltage larger than recharger voltage, we treated
as charging timeout, of course in this case will stop charging, and we
will restart charge if meet stop charger for more than 3hours.

Change-Id: I97b9efb4d0e500d6f0c0173b5c1587b0e275adb5
Signed-off-by: Yuanjiang Yu <yuanjiang.yu@spreadtrum.com>