Bug #1055406 power: supply: fan54015: Fix kernel panic issue when phone reboot test
[root cause  ] fix kernel panic issue when phone reboot test
[changes     ] fix kernel panic issue when phone reboot test
[side effects] no
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: Sharkl3
[test case]: compile ok
[reviewers   ] baolin.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

When the phone boot process, becauese the fan54015 charger depend on phy
registration, so if the phy resource is not ready, the fan54015 charger will
always probe, and the bus regulator will register repeatedly but in this
moment the debtry->d_inode may be 0, and currently the musb will access
the vbus regulator interface, so will led kernel panic. Now we register
vbus regulator after fan54015 charger get phy handle.

Change-Id: I6c3dc89c5ce220d89990e18d5fd9300ddc9e3b53
Signed-off-by: Yuanjiang Yu <yuanjiang.yu@spreadtrum.com>