Bug #1005477 sched: fair: fix the value of cpu_capacity_orig
[root cause  ] cpu_capacity_orig is unreasonable
[changes     ] fix the value of cpu_capacity_orig
[side effects]no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:yes
[board]:sharkl5
[test case]:scheduler
[reviewers   ] ke wang
[change_type ] feature_bugfix
[tag_product ] common

1.when IPA is working,max_cpu_capacity of root domain will be less than
cpu_capacity_orig.it is correct that the max_cpu_capacity of root domain
is greater or equal to cpu_capacity_orig.it may cause load balance or
select cpu to be unreasonable.
2.cpu_capacity's is the sum of cfs task util,when IPA is working,
cpu_capacity will be reduced,check_cpu_capacity function will make
it easier to return a value of true,this will make system easier to 
migrate task,maybe the task can run reasonably.

Change-Id: Idd39b5ec32d11007d39ffffcee55e334f15903b7
Signed-off-by: samer.xie <samer.xie@unisoc.com>