Bug #1003594 power: supply: fan54015: Modify fan54015 otg restart strategy
[root cause  ] modify fan54015 otg restart strategy
[changes     ] modify fan54015 otg restart strategy
[side effects] no
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: ums312 2h10
[test case]: compile ok
[reviewers   ] baolin.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

On this platform we use a GPIO to detect if the OTG is abnormal. But USB
will also use the GPIO as an extcon event to notify the USB whether
there is a device plug. This causes the USB and the charger to apply for
a GPIO at the same time, which causes the charger request to fail. So
now we disable the OTG stable validation feature to avoid this issue,
and we should find a good solution to solve this issue in future.

Change-Id: I9e3df55319780e90668f16bdbc87ad247473323c
Signed-off-by: Yuanjiang Yu <yuanjiang.yu@spreadtrum.com>