Bug #1060439 power: supply: sc2703: Fix flash high probability not flash issue
[root cause  ] fix flash high probability not flash issue
[changes     ] fix flash high probability not flash issue
[side effects] no
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: Sharkl5
[test case]: compile ok, charger ok
[reviewers   ] baolin.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Due to the fact that the flashing light will be in a dark environment,
there will be a probability that it will not flash. We set the VCENT
voltage to 4.8v, which can improve the efficiency of flash.

Change-Id: I00bb4f808b2b41c39718b88b400d5d31537ace6e
Signed-off-by: Yuanjiang Yu <yuanjiang.yu@spreadtrum.com>
(cherry picked from commit c2375b4ebd52e0cd792e85cd7061121cab5ce983)