Bug #1046536 WORKAROUND: tty: serial_core: Fix the console serial problems at serial port resume
[root cause ] there are errors such as baud rate and date length of console serial at serial port resume.
[changes ]
[side effects]
[self test ]: yes
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: yes
[confirm dependent commit]: yes
[board]: sharkl3
[test case]: power on is ok
[reviewers ]
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

On the UNISOC SOC, in some special scenarios, userspace needs to close console.
But console logs are still needed in the kernel. When console resume, console needs
to configure the baud rate and data length of the serial port according to its
own cflag. At this time, the cflag is 0, which will cause serial port to produce
configuration errors that do not meet user expectations.

When userspace opens a serial port for console, uart_port_startup() is called. This
function assigns the uport->cons->cflag value to TTY->termios.c_cflag, then it is
cleared to 0. When the user space closes this serial port, the TTY structure will
be released, and At this time uport->cons->cflag has also been cleared at port open.
It will cause the cflag get value is 0 at console serial resume from system deep.
But the console serial also use this cflag to configure the baud rate and data length.
It will cause serial port to produce configuration errors.

To fix this, Assigns the TTY->termios.c_cflag value to uport->cons->cflag before the user
space closes this console serial port. It will ensure that the correct cflag value can be
gotten when the console serial port was resumed.

Change-Id: I9290ef120ff75d3bb80993b5b663c4f39d6ed059
Signed-off-by: Lanqing Liu <lanqing.liu@spreadtrum.com>