Bug #1072063 PCI: dwc: Support PCIe wakeup function
[root cause  ] PCIe EP needs to wakeup AP
[changes     ] pci driver
[side effects] no
[self test   ]: yes
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: no
[is there dependence]: no
[confirm dependent commit]: yes
[board]: ud710-3h10
[test case]: PCIe works well
[reviewers   ] Jason Wu, Jolly Li
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

When AP enter deep, an endpoint(e.g. marlin3 or orca) may doesn't sleep.
If the endpoint wants to transmit data, it must wakeup AP by wake#
signal. The wakeup pin from EP is connected to EIC_EXT1 which We can
think of as a gpio (although it isn't). In order to support wakeup
function, this signal must be converted to a irq.

Change-Id: I3977219098efa067934782c05fa187bb5b61aa21
Signed-off-by: Billows Wu <Billows.Wu@spreadtrum.com>