Bug #1068543 arm64: dts: Modify vddwcn status to always on for orca.
[root cause  ] Modify vddwcn status to always on for orca.
[changes     ] Modify vddwcn status to always on for orca.
[side effects] no
[self test   ]: yes
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: yes
[board]: udx710_2h10
[test case]: normal boot
[reviewers   ] erick.chen
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

In orca version 1.0.0 vddwcn supply power for 234G Lajolla VDD0V9DIG and
in orca version 1.0.1 vddwcn is not used. Meanwhile the project require the
version 1.0.1 code can use for version 1.0.0, So to make sure the 1.0.0 work
normal,we modify vddwcn status to always on and later we will make distinction
in the uboot.

Change-Id: I867f4d2a20925569324a73f6df0c9584fb0198f2
Signed-off-by: Sherry Zong <sherry.zong@unisoc.com>