Bug #1008857 proc/sched: Add /proc/$PID/schedinfo interface
[root cause  ] match the upper layer boost/unboost policy
[changes     ] add s_sum_exec_runntime and b_sum_exec_runntime in sched
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:yes
[board]:sharkL5
[test case]: check /proc/$PID/schedinfo after boot
[reviewers   ]
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Add /proc/$PID/schedinfo interface, which will show not only the
sum_exec_runtime of the thread but also sum_exec_runtime of the whole
process in which can have more than one thread. There are 5 types
sum_exec_runntime: runtime of this thread on small cores, runtime
of this thread on big cores, runtime of the process on small cores,
runtime of the process on big cores and runtime of the process on
all cores.

Change-Id: I1ac768492c83f452b7318f3387520a42c1b919e0
Signed-off-by: Ke Wang <ke.wang@spreadtrum.com>