Bug #1028350 WORKAROUND: PCI: sprd: Disabled MSI 0~3 in MSI controller
[root cause  ] hw interrupt workaround
[changes     ] pci driver
[side effects] no
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: ud710_2h10
[test case]: boot
[reviewers   ] billows.wu; jason.wu
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Orca/Roc1 pci asic design for normal interrupt mode is not
compatible with kernel PCIE framework, the 0-3 msi irqs are handled
directly by hardware. This workaround is used to disable the
corresponding bits of these four irqs on the msi reqister so that AP GIC
will not receive these interrupts.

Change-Id: Ice6076b7adb93105d44aed4bd68a93dfe9362f46
Signed-off-by: Jolly Li <jolly.li@spreadtrum.com>