Bug #1075033 arm64: dts: sprd: correct sd card  EIC pin number
[root cause  ] correct the sd card EIC pin number
[changes     ] ud710-2h10.dts
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:no
[is there dependence]:no
[confirm dependent commit]:no
[board]:roc1+orca
[test case]:normal test
[reviewers   ] Billows.Wu
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

SD Card detect pin are connected to EIC_EXT4, In EIC_EXT4, we can
determine the sd card pin number through the EIC_source_list file. In
this file, It recoard the bit17~bit19 in the address 0x3243c030
(PIN_CTRL_REG5) to distinguish which card is used for card plug-in/out
detect select.

when bit19~bit17 value is 000, it means we use the sd0 card.
when bit19~bit17 value is 001, it means we use the sd1 card.
......
In roc1 project, we only use the sd0 card for card plug-in/out detect
function. So the pin number is 3. Therefore,we can calculate the
eic_async value is (8*4+3) not 19.

Change-Id: I36f13995fb07d88d17c8f02386f32ab125d39f9f
Signed-off-by: Xuefei Wang <xuefei.wang@unisoc.com>