Bug #1066010 power: supply: charger-manager: Fix the full charger stop issue
[root cause  ] fix the full charger stop issue
[changes     ] fix the full charger stop issue
[side effects] no
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: Roc1
[test case]: compile ok
[reviewers   ] baolin.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

When met the charging cutoff condition, we will stop charging. when the battery
voltage slowly decrease, if met the restart charging condition, it will restart
charging. But now the issue is not met the restart charging condition, already
start charging, because the charger manage monitor workqueue check charge
voltage is meet the restarting conditions. so we check charging fault function,
when only the fault is disappear, we will restart charging.

Change-Id: I3baeac593bf34b2b4c20ae7dd6169610a752fe27
Signed-off-by: Yuanjiang Yu <yuanjiang.yu@spreadtrum.com>