Bug #1090665 staging: sprd: sipa: sipa eth deal with mac header distinguishly
[root cause  ] sipa_eth receive a skb with a mac header on ORCA only platform,
but has no mac header on ROC1+ORCA platform.
[changes     ] add mac-header bool in dts to distinguish,
if sipa_eth receive a skb with a mac header,
we have to pull this header since we use rawip scheme
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:ud710_2h10
[test case]:ping 8.8.8.8, it works
[reviewers   ] junjie.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Sipa_eth receive a skb with a mac header on ORCA only platform,
but has no mac header on ROC1+ORCA platform,
so we have to add mac-header bool in dts to distinguish,
if sipa_eth receive a skb with a mac header,
we have to pull this header since we use rawip scheme
This patch distinguishly to deal with mac header
on orca cpe & roc1+orca platform.

Change-Id: Id24689ebae8a7b40fe747acc330a8e5c9fdb8136
Signed-off-by: Wade Shu <wade.shu@unisoc.com>