Bug #1071859 drm/sprd: limit gsp bandwidth output capability
[root cause  ] dpu underflow
[changes     ] sprd drm driver
[side effects] none
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:ums312_HLTE217T
[test case]:bandwidth test ok
[reviewers   ] leon.he thomas.chai
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Gsp generates very large bandwidth when writing data to memory
If you don't limit it, it will been completed in 1-3ms.
This will have a significant impact on the entire ddr bandwidth
resources, the most direct consequences will lead to dpu underflow.

We limit gsp to write data in a vsync cycle(16ms), not 1-3ms.
so that the instantaneous bandwidth is averaged. But this may affect
the WRITEBACK(WB) write back data to ddr, so need to find an optimal
balance between limiting gsp output capability and writeback.

Change-Id: I73f6ec25f87a49bf4bfbfe91ad109cb0ce4bb00c
Signed-off-by: Kevin Tang <kevin.tang@spreadtrum.com>