Bug #1093444 power: supply: sc2703: Modify sc2703 to stop charging control mode
[root cause  ] Modify 2703 to stop charging control mode
[changes     ] Modify 2703 to stop charging control mode
[side effects] no
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: Sharkl5
[test case]: compile ok
[reviewers   ] baolin.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Now we turn off the sc2703 charger buck to stop charging, this way will
lose the ability to power the phone system. So we trun off the sc2703
charge enable to replace it.In the production line calibration mode,
some test items require the phone to be completely powered down and then
tested, so it needs to be turned off sc2703 charger buck.

Change-Id: I41e8a73b46567fc569a4bb889a0c7f34b637fed4
Signed-off-by: Yuanjiang Yu <yuanjiang.yu@spreadtrum.com>