Bug #1075477 drm/sprd: limit gsp bandwidth output capability for roc1
[root cause  ] dpu underflow
[changes     ] sprd drm driver
[side effects] none
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:roc1_3h10
[test case]:SR switch ok
[reviewers   ] chen.he leon.he
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Gsp generates very large bandwidth when writing data to memory
If you don't limit it, it will been completed in 1-3ms.
This will have a significant impact on the entire ddr bandwidth
resources, the most direct consequences will lead to dpu underflow.

We limit gsp to write data in a vsync cycle(16ms), not 1-3ms.
so that the instantaneous bandwidth is averaged. But this may affect
the WRITEBACK(WB) write back data to ddr, so need to find an optimal
balance between limiting gsp output capability and writeback.

Change-Id: I9377db780cd7294445962a24eacacef8d4463973
Signed-off-by: Kevin Tang <kevin.tang@spreadtrum.com>