Bug #1058871 staging: ipa: add delay after hash table switch
[root cause  ] add delay after hash table switch
[changes     ] ipa
[side effects] no
[self test   ]: ipa
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: no
[is there dependence]: no
[confirm dependent commit]: yes
[board]: roc1
[test case]: ipa
[reviewers   ] fei.zhang zhiwei.wu alvin.zhou
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

When the IPA routing table cuts the table, it does not wait until the
routing table actually writes the DDR, and returns the done. At this
time, the routing table obtained by the IPA may not be the latest,
and only the timestamp field in the routing table is included. Through
the 500us delay, can guarantee the IPA to get the latest table at a high
probability.

Change-Id: Ia9fd0f3074a9e42d2f18ebbf6a275272015d1ec9
Signed-off-by: Qingsheng Li <qingsheng.li@spreadtrum.com>