Bug #1099147 mmc: sprd-sdhc: mmc host driver requests probe retry
[root cause  ] get the mmc regulator resource fail will cause boot fail
[changes     ] sprd-sdhcr10.c
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:no
[is there dependence]:no
[confirm dependent commit]:no
[board]:pike2
[test case]:normal test
[reviewers   ] Billows.Wu
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

In the regulator driver, some regulators need to be disabled with the
use count value of zero, it means those regulators can not supply power
any more. MMC host driver gets the regulator resource failed, and we
ignore this fail. This makes the regulator use count to be zero for sdio
devices, and the sdio devices power down.

Change-Id: I942f9648c9f0be309d8ce0bc9f0e3594e3a6bd88
Signed-off-by: Xuefei Wang <xuefei.wang@unisoc.com>