Bug #926329 usb: gadget: u_ether: notify sipa_nic about the real flowctrl timing
[root cause  ] network subsystem fail to be notified about the event
that exited a flow control, resulting in the network subsystem
cannot recover after enter flow ctrl
[changes     ] we notify sipa_nic about that the time 
we really stop queueing when we start do flowctl
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:ud710_2h10
[test case]:okay to invoke this interface, flowctrl mechanism works fine
[reviewers   ] junjie.wang, qingsheng.li
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Network subsystem fails to be notifyed about the event that exiting a flow control,
resulting in a failure of recoving after a flow ctrl.
This patch notify sipa_nic about the real timing that we stop queueing.

Change-Id: Ia11b0cb5a99e86372b217d02bf7a69d089035d61
Signed-off-by: Wade Shu <wade.shu@unisoc.com>