Bug #1101872 arm64: dts: Modify restart charging threshold after battery full for ums312 2h10
[root cause  ] Modify restart charging threshold after battery full for ums312 2h10
[changes     ] Modify restart charging threshold after battery full for ums312 2h10
[side effects] no
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: Sharkl5-2
[test case]: compile ok
[reviewers   ] baolin.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Due to the fuel gauge accuracy and fan54015 cccv accuracy have 30000uv.
The project use 4350000uv battery, so we set the cut off voltage as 4320000uv,
but if the restart charging threshold set to 60000uv. Now the issue is after
charging end, restart charging until the battery capacity show 95%.
So we should set restart charging threshod to 20000uv. We want to restart
charging as the battery capacity show about 98%.

Change-Id: I466e84f36b69cd9c925c5c3f814adeb9cf4dec4c
Signed-off-by: Yuanjiang Yu <yuanjiang.yu@spreadtrum.com>