Bug #1097108 power: supply: charger-manager: Fix the inital battery capacity is not accurate issue
[root cause  ] Fix the inital battery capacity is not accurate issue
[changes     ] Fix the inital battery capacity is not accurate issue
[side effects] no
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: Sharkl5
[test case]: compile ok
[reviewers   ] baolin.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Now we find that get the battery capacity is not accurate in phone boot.
Analyze the reason may be the last battery capacity not saved in shutdown
process, so read the battery capacity is not accurate after phone boot.
Now we improvement action is save the battery capacity in the rtc register
when the battery capacity update.

Change-Id: Iaec8ea061b820dd2bfa80f2249d644d2721c27b7
Signed-off-by: Yuanjiang Yu <yuanjiang.yu@spreadtrum.com>