Bug #1117526 staging: sprd: mailbox: The irq of non-waking up outbox must remove no suspend flag
[root cause  ] The irq of none wakeup box must remove no suspend flag.
[changes     ] mailbox driver files
[side effects] none
[self test   ]:9863a1h10 AP can goto sleep when sensor hub is working.
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:no
[is there dependence]:no
[confirm dependent commit]:no
[board]:9863a1h10
[test case]:AP can deep sleep.
[reviewers   ] fei.zhang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

Because the sensorhub processor always send some data to ap frequently,
and this data that does not need to wakeup the ap and will send to 
the none wakeup box, so must remove the no suspend flag for the irq of 
non-waking up outbox. Otherwise, it may prevent system suspend.

Change-Id: I98a938c8f0de25f11980e107df79979f30adef26
Signed-off-by: Wenping Zhou <wenping.zhou@spreadtrum.com>