Bug #1110626 arm64: dts: Remove unused GPIO node
[root cause  ] Cancel detection eic0 to check otg function is normal
[changes     ] Cancel detection eic0 to check otg function is normal
[side effects] no
[self test   ]:
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: no
[board]: Sharkl5_2h10
[test case]: compile ok
[reviewers   ] baolin.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

On this platform we use a GPIO to detect if the OTG is abnormal. But USB
will also use the GPIO as an extcon event to notify the USB whether
there is a device plug. This causes the USB and the charger to apply for
a GPIO at the same time, which causes the charger request to fail. So we
will enable otg restart workqueue to enable otg funciton to workaround this
issue.

Change-Id: I5d6a0c4d33174cb5b42ae27ef33e50a3f79e765d
Signed-off-by: Yuanjiang Yu <yuanjiang.yu@unisoc.com>