Bug #1114469 arm64: dts: Change cm-fullbatt-voltage for sharkl5pro
[root cause  ] change cm-fullbatt-voltage
[changes     ] change cm-fullbatt-voltage
[side effects] no
[self test   ]:yes
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: yes
[board]: sharkl5pro
[test case]: compile ok
[reviewers   ] yuanjiang.yu
[change_type ] dts config
[tag_product ] common

This issue is due to the charge IC fan54015 error, and hardware engineer 
make a test, give the test data, and suggest software engineer adjust the
fullbatt threshold to cover the fan54015 error 1%, then the fullbatt 
threshold is under 4306. Disscussing with Yuanjiang Yu, finally, we set
the fullbatt threshold to 4300.

Change-Id: I6ffd23e76b0bae86f63d4d8e22a38f590fcb42ed
Signed-off-by: Rong Wu <rong.wu@unisoc.com>