Bug #1119943 net: add a sync member to address a seq-synchronization issue
[root cause  ] when doing esp seq-synchronization, 
if there is a esp pkt which is unrelated to lowpower scheme, 
is send out, it will set imsbr_cp_esp_sync to 0, 
then the later synchroniztion will fail to get a right sequence from cp.
[changes     ] add a sync flag for each esp flow, 
so we always synchronize from cp for each esp flow successfully.
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:ums312_1h10
[test case]:vowifi lowpower call success
[reviewers   ] junjie.wang
[change_type ] bringup --> sprd_chip_support
[tag_product ] common

We need a sync flag for each esp flow, 
otherwise it is hard to determine which esp flow need a synchroniztion from cp esp info.

Change-Id: I71ed1c7a5ceb067f2db729c02d284db9b82f59ad
Signed-off-by: Wade Shu <wade.shu@unisoc.com>